[chore/#127-menu-directory-refactor]
