+++ src/main/java/yuseteam/mealticketsystemwas/domain/order/service/OrderService.java
@@ -3,8 +3,8 @@ package yuseteam.mealticketsystemwas.domain.order.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import yuseteam.mealticketsystemwas.domain.menu.entity.Menu;
-import yuseteam.mealticketsystemwas.domain.menu.repository.MenuRepository;
+import yuseteam.mealticketsystemwas.domain.menu.common.entity.Menu;
+import yuseteam.mealticketsystemwas.domain.menu.common.repository.MenuRepository;
 import yuseteam.mealticketsystemwas.domain.oauthjwt.entity.User;
 import yuseteam.mealticketsystemwas.domain.oauthjwt.repository.UserRepository;
 import yuseteam.mealticketsystemwas.domain.order.dto.CreateOrderReq;
